ref(utils): move generic function to generic file
